Upgrade to ruby-2.6.6 (#70)
At some point when we re-built the base image we seem to have grabbed an updated version of phusion/passenger-ruby26, which included an upgrade from ruby-2.6.5 to ruby-2.6.6 and now we're getting a ruby version mis-match. Let's upgrade our application to ruby-2.6.6 to match the expectations of the base image.