Dockerfile Consolidation
After playing with multi-stage Docker builds, I no longer think it is exactly what we need for this project. We are not dealing with build dependencies that we do not want in the final build, but are instead working with trying to get a stable platform that is as uniform and efficient as possible for each project to work from. That let me to feel like there was a single base that both projects should pull from. It still makes each project simpler to reason about, while reducing the deploy footprint. It does not force both projects to upgrade to a new base at the same time, but any time they are in sync than all layers of the base will be shared between the two projects which will aid in disk space / network up and down speed.

This PR comes with a sister PR over in Camerata which is where I put the Base (yalelibrary/yul-dc-camerata#44) . I not sure that's the right spot, but it didn't seem like a terrible spot to start with.