Implement security for IIIF images
This commit adds a new route check-iiif which tests for authorization in whatever way other items test for authorization.
It also adds Nginx configs to allow the Cantaloupe service to be moved to authorized-iiif and for this app to server /iiif.

All upstream image transfer after authorization would then be streamed by nginx and would free the Rails application process.

See https://medium.com/@jefawks3/using-nginx-as-a-multi-step-reverse-proxy-to-authorize-requests-before-serving-content-on-s3-d4094577c60e